+++ dphonebook/lib/providers/__init__.py
@@ -1,6 +1,7 @@
 from dphonebook.lib.providers.dummy_example_com import DummyExampleCom
 from dphonebook.lib.providers.receive_sms_co import ReceiveSmsCo
 from dphonebook.lib.providers.receive_sms_online_info import ReceiveSmsOnlineInfo
+from dphonebook.lib.providers.receive_sms_org import ReceiveSmsOrg
 from dphonebook.lib.providers.receive_smss import ReceiveSmss
 
 
@@ -8,5 +9,6 @@ number_provider_classes = [
     DummyExampleCom,
     ReceiveSmss,
     ReceiveSmsCo,
-    ReceiveSmsOnlineInfo
+    ReceiveSmsOnlineInfo,
+    ReceiveSmsOrg
 ]